chore: Add a stub for benchmarking
