+++ webpack.config.js
@@ -1,128 +1,18 @@
 const path = require("path");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
-const CpuProfilerWebpackPlugin = require("cpuprofile-webpack-plugin");
 const { WebpackPluginServe } = require("webpack-plugin-serve");
 const { merge } = require("webpack-merge");
 const { target } = require("webpack-nano/argv");
+const parts = require("./webpack.parts");
 
 const project = process.env.PROJECT || "template";
 const builder = process.env.BUILDER || "esbuild";
-
-const builderAlternatives = {
-  esbuild: {
-    test: /\.([t|j]sx?|svg)$/,
-    loader: "esbuild-loader",
-    exclude: /node_modules/,
-    options: {
-      loader: "tsx",
-      target: "es2015",
-    },
-  },
-  swc: {
-    test: /\.([t|j]sx?)$/,
-    loader: "swc-loader",
-    exclude: /node_modules/,
-    options: {
-      jsc: {
-        parser: {
-          syntax: "typescript",
-        },
-      },
-    },
-  },
-  babel: {
-    test: /\.([t|j]sx?)$/,
-    loader: "babel-loader",
-    exclude: /node_modules/,
-    options: {
-      presets: [
-        ["@babel/preset-env", { targets: "defaults" }],
-        "@babel/preset-typescript",
-        "@babel/preset-react",
-      ],
-    },
-  },
-};
-
-const projects = {
-  chromatic: {
-    module: {
-      rules: [
-        {
-          test: /\.(m?[t|j]s)$/,
-          resolve: {
-            fullySpecified: false,
-          },
-        },
-        {
-          test: /\.svg$/,
-          loader: "react-svg-loader",
-          options: {
-            jsx: true,
-          },
-        },
-        {
-          test: /\.(graphql|gql)$/,
-          include: [/schema/],
-          exclude: /node_modules/,
-          loader: "raw-loader",
-        },
-        {
-          test: /\.handlebars/,
-          loader: "handlebars-loader",
-          exclude: /node_modules/,
-          // query: {
-          //   helperDirs: path.join(__dirname, '..','lib', 'emails', 'helpers'),
-          // },
-        },
-      ],
-    },
-  },
-  "design-system": {
-    module: {
-      rules: [
-        {
-          test: /\.m?[t|j]sx?$/,
-          resolve: {
-            fullySpecified: false,
-          },
-        },
-        {
-          test: /\.css$/,
-          use: ["style-loader", "css-loader"],
-        },
-        {
-          test: /\.(png|jpe?g|gif|svg)$/i,
-          use: [
-            {
-              loader: "file-loader",
-            },
-          ],
-        },
-      ],
-    },
-  },
-  template: {
-    module: {
-      rules: [
-        {
-          test: /\.css$/,
-          use: ["style-loader", "css-loader"],
-        },
-      ],
-    },
-  },
-};
-
-const cpuProfiler = {
-  plugins: [new CpuProfilerWebpackPlugin()],
-};
 
 const commonConfig = merge(
   {
     plugins: [new HtmlWebpackPlugin({ filename: "iframe.html" })],
     module: {
-      rules: [builderAlternatives[builder]],
+      rules: [parts.builderAlternatives[builder]],
     },
     resolve: {
       extensions: [".ts", ".tsx", ".js", ".jsx", ".json"],
@@ -137,7 +27,7 @@ const commonConfig = merge(
       path: path.resolve(__dirname, "dist"),
     },
   },
-  projects[project],
+  parts.projects[project],
   process.env.COMPILE_LAZILY === "1"
     ? {
         experiments: {
@@ -145,7 +35,7 @@ const commonConfig = merge(
         },
       }
     : {},
-  process.env.PROFILE_CPU === "1" ? cpuProfiler : {}
+  process.env.PROFILE_CPU === "1" ? parts.cpuProfiler : {}
 );
 
 const developmentConfig = {
