+++ webpack.parts.js
@@ -1,0 +1,117 @@
+const CpuProfilerWebpackPlugin = require("cpuprofile-webpack-plugin");
+
+const builderAlternatives = {
+  esbuild: {
+    test: /\.([t|j]sx?|svg)$/,
+    loader: "esbuild-loader",
+    exclude: /node_modules/,
+    options: {
+      loader: "tsx",
+      target: "es2015",
+    },
+  },
+  swc: {
+    test: /\.([t|j]sx?)$/,
+    loader: "swc-loader",
+    exclude: /node_modules/,
+    options: {
+      jsc: {
+        parser: {
+          syntax: "typescript",
+        },
+      },
+    },
+  },
+  babel: {
+    test: /\.([t|j]sx?)$/,
+    loader: "babel-loader",
+    exclude: /node_modules/,
+    options: {
+      presets: [
+        ["@babel/preset-env", { targets: "defaults" }],
+        "@babel/preset-typescript",
+        "@babel/preset-react",
+      ],
+    },
+  },
+};
+
+const projects = {
+  chromatic: {
+    module: {
+      rules: [
+        {
+          test: /\.(m?[t|j]s)$/,
+          resolve: {
+            fullySpecified: false,
+          },
+        },
+        {
+          test: /\.svg$/,
+          loader: "react-svg-loader",
+          options: {
+            jsx: true,
+          },
+        },
+        {
+          test: /\.(graphql|gql)$/,
+          include: [/schema/],
+          exclude: /node_modules/,
+          loader: "raw-loader",
+        },
+        {
+          test: /\.handlebars/,
+          loader: "handlebars-loader",
+          exclude: /node_modules/,
+          // query: {
+          //   helperDirs: path.join(__dirname, '..','lib', 'emails', 'helpers'),
+          // },
+        },
+      ],
+    },
+  },
+  "design-system": {
+    module: {
+      rules: [
+        {
+          test: /\.m?[t|j]sx?$/,
+          resolve: {
+            fullySpecified: false,
+          },
+        },
+        {
+          test: /\.css$/,
+          use: ["style-loader", "css-loader"],
+        },
+        {
+          test: /\.(png|jpe?g|gif|svg)$/i,
+          use: [
+            {
+              loader: "file-loader",
+            },
+          ],
+        },
+      ],
+    },
+  },
+  template: {
+    module: {
+      rules: [
+        {
+          test: /\.css$/,
+          use: ["style-loader", "css-loader"],
+        },
+      ],
+    },
+  },
+};
+
+const cpuProfiler = {
+  plugins: [new CpuProfilerWebpackPlugin()],
+};
+
+module.exports = {
+  builderAlternatives,
+  projects,
+  cpuProfiler,
+};
